moved serde_api to own module mostly synced up with altrios added tests for resources and made resources more robust
