+++ fastsim-core/resources/interpolators/res_default_pwr.yaml
