+++ fastsim-core/resources/interpolators/res_default_pwr_and_temp.yaml
