+++ fastsim-core/resources/interpolators/res_default_pwr_soc_and_temp.yaml
