+++ fastsim-core/src/utils/interp.rs
@@ -114,5 +114,23 @@ impl Init for Interpolator {
 }
 impl SerdeAPI for Interpolator {
     #[cfg(feature = "resources")]
-    const RESOURCE_PREFIX: &'static str = "interpolators";
+    const RESOURCE_SUBDIR: &'static str = "interpolators";
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    type StructWithResources = Interpolator;
+
+    #[test]
+    fn test_resources() {
+        let resource_list = StructWithResources::list_resources();
+        assert!(!resource_list.is_empty());
+
+        // verify that resources can all load
+        for resource in resource_list {
+            StructWithResources::from_resource(resource, false).unwrap();
+        }
+    }
 }
