+++ fastsim-core/src/vehicle/vehicle_model.rs
@@ -1,6 +1,4 @@
 use crate::prelude::*;
-#[cfg(feature = "resources")]
-use crate::resources;
 
 use super::{hev::HEVPowertrainControls, *};
 pub mod fastsim2_interface;
@@ -127,7 +125,7 @@ impl Vehicle {
     #[staticmethod]
     /// list available vehicle resources
     fn list_resources_py() -> Vec<String> {
-        resources::list_resources(Self::RESOURCE_PREFIX)
+        Self::list_resources()
     }
 
     /// Load vehicle from file saved in fastsim-2 format
@@ -230,7 +228,7 @@ impl Mass for Vehicle {
 
 impl SerdeAPI for Vehicle {
     #[cfg(feature = "resources")]
-    const RESOURCE_PREFIX: &'static str = "vehicles";
+    const RESOURCE_SUBDIR: &'static str = "vehicles";
 }
 impl Init for Vehicle {
     fn init(&mut self) -> Result<(), Error> {
@@ -826,5 +824,18 @@ pub(crate) mod tests {
         let sd = crate::simdrive::SimDrive::new(veh, cyc, Default::default());
         let mut sd2 = sd.to_fastsim2().unwrap();
         sd2.sim_drive(None, None).unwrap();
+    }
+
+    type StructWithResources = Vehicle;
+
+    #[test]
+    fn test_resources() {
+        let resource_list = StructWithResources::list_resources();
+        assert!(!resource_list.is_empty());
+
+        // verify that resources can all load
+        for resource in resource_list {
+            StructWithResources::from_resource(resource, false).unwrap();
+        }
     }
 }
